Because ingredient and condiment is practically the same object we join booth class to one class called Product
