+++ src/main/java/com/overlordsystems/sanguchetto/ingredient/Product.java
@@ -1,0 +1,79 @@
+package com.overlordsystems.sanguchetto.ingredient;
+
+import java.util.Objects;
+
+public class Product {
+    private int id;
+    private String name;
+    private int stock;
+    private double price;
+    private TypeOfProduct type;
+
+    public Product(){
+
+    }
+
+    public Product(int id, String name, int stock, double price, TypeOfProduct type){
+        this.id = id;
+        this.name = name;
+        this.stock = stock;
+        this.price = price;
+        this.type = type;
+    }
+
+    public Product(int id){
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public int getStock() {
+        return stock;
+    }
+
+    public void setStock(int stock) {
+        this.stock = stock;
+    }
+
+    public double getPrice() {
+        return price;
+    }
+
+    public void setPrice(double price) {
+        this.price = price;
+    }
+
+    public TypeOfProduct getType() {
+        return type;
+    }
+
+    public void setType(TypeOfProduct type) {
+        this.type = type;
+    }
+
+    public int getId() {
+        return id;
+    }
+
+    public void setId(int id){
+        this.id = id;
+    }
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        Product product = (Product) o;
+        return id == product.id;
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(id);
+    }
+}
